pyproject.toml
[tool.poetry]
name = "django-collabconnect"
version = "0.1.0"
description = ""
authors = ["Aditya Pratap Singh <adityapratapsingh51@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.2"
Django = "^3.2.6"
django-heroku = "^0.3.1"
gunicorn = "^20.1.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.8.0"
dnspython = "^2.1.0"
psycopg2 = "^2.9.1"
dj-database-url = "^0.5.0"
pymongo = "^3.12.0"
cryptography = "^3.4.7"
google-auth = "^2.0.0"
requests = "^2.26.0"
pycryptodomex = "^3.10.1"
python-dotenv = "^0.19.0"
urllib3 = "^1.26.6"
PyJWT = "^2.1.0"
firebase-admin = "^5.1.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
